+++ example/app/_deps.js
@@ -1,2 +1,2 @@
-var bldr = require('../..')('ex', __filename);
-bldr.browser('/vendor/jquery-2.1.1');
+var loadstar = require('../..')('ex', __filename);
+loadstar.browser('/vendor/jquery-2.1.1');