+++ example/app/admin/init.js
@@ -1,5 +1,5 @@
 var init = function() {
-  var user = new ex.models.User({name: 'bldr-admin'});
+  var user = new ex.models.User({name: 'loadstar-admin'});
   var dash= new ex.admin.Dashboard({model: user, el: document.getElementById('app')});
   dash.render();
 }