+++ example/app/index.js
@@ -1,7 +1,7 @@
-var bldr = require('../..')('ex', __filename, {appDir: __dirname});
+var loadstar = require('../..')('ex', __filename, {appDir: __dirname});
 
 require('./_deps');
 require('./models');
-bldr.define('views/Login');
-bldr.browser('./init');
-module.exports = bldr.ex;
+loadstar.define('views/Login');
+loadstar.browser('./init');
+module.exports = loadstar.ex;