+++ test/expects/should-only-define-shared-keys-once-test-two.js
@@ -1,17 +1,17 @@
 ;(function() {
 
 // exportmore.js
-testbldr = {};
-testbldr.fixtures = {};
-testbldr.fixtures.exportmore = {
+testloadstar = {};
+testloadstar.fixtures = {};
+testloadstar.fixtures.exportmore = {
   msg: function() {
     return 'more';
   }
 };
 
 // exports.js
-testbldr.fixtures.exports = {};
-testbldr.fixtures.exports.msg = function() {
+testloadstar.fixtures.exports = {};
+testloadstar.fixtures.exports.msg = function() {
 };
 
 })()