+++ test/expects/should-replace-module.exports-if-define-is-used.js
@@ -1,9 +1,9 @@
 ;(function() {
 
 // exports.js
-testbldr = {};
-testbldr.exports = {};
-testbldr.exports.msg = function() {
+testloadstar = {};
+testloadstar.exports = {};
+testloadstar.exports.msg = function() {
 };
 
 })()