+++ test/expects/should-wrap-with-amd-and-define.js
@@ -1,9 +1,9 @@
 (function(root,factory) { if (typeof define === "function" && define.amd) {   define([], factory); } else { root.testglobal = factory(); }})(this, function() {
 
 // exports.js
-var testbldr = {};
-testbldr.exports = {};
-testbldr.exports.msg = function() {
+var testloadstar = {};
+testloadstar.exports = {};
+testloadstar.exports.msg = function() {
 };
 
 return testglobal;