+++ test/fixtures/a.js
@@ -1,4 +1,4 @@
-var bldr = require('../..')('testbldr', __filename);
-var b = bldr.require('./b');
+var loadstar = require('../..')('testloadstar', __filename);
+var b = loadstar.require('./b');
 exports.msg = 'a';
 exports.b = b;