+++ test/fixtures/extends.js
@@ -1,5 +1,5 @@
-var bldr = require('../..')('testbldr', __filename);
-var testglobal = bldr.require('./testglobal');
+var loadstar = require('../..')('testloadstar', __filename);
+var testglobal = loadstar.require('./testglobal');
 
 testglobal.msg = function() {
 };